fold GitHubRun into github launcher
