+++ src/discord-cluster-manager/bot.py
@@ -4,6 +4,7 @@ import os
 
 import consts
 import discord
+import env
 import uvicorn
 from api.main import app, init_api
 from cogs.admin_cog import AdminCog
@@ -78,7 +79,7 @@ class ClusterBot(commands.Bot):
             # Load cogs
             submit_cog = SubmitCog(self)
             submit_cog.register_launcher(ModalLauncher(consts.MODAL_CUDA_INCLUDE_DIRS))
-            submit_cog.register_launcher(GitHubLauncher())
+            submit_cog.register_launcher(GitHubLauncher(env.GITHUB_REPO, env.GITHUB_TOKEN))
             await self.add_cog(submit_cog)
             await self.add_cog(BotManagerCog(self))
             await self.add_cog(LeaderboardCog(self))
